requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
#greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
jsonify==0.5
MarkupSafe==2.0.1
marshmallow==3.12.1
marshmallow-sqlalchemy==0.26.1
#ngrok==0.0.1
#psycopg2
#psycopg2-binary==2.8.6
PyJWT
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.18
twilio
urllib3==1.26.5
Werkzeug==2.0.1
validators==0.20.0
